+++ HISTORY.md
@@ -16,6 +16,7 @@
 * Remove deprecated API DBOption::base_background_compactions.
 * Remove deprecated API DBOptions::purge_redundant_kvs_while_flush.
 * Remove deprecated overloads of API DB::CompactRange.
+* Remove deprecated option DBOptions::skip_log_error_on_recovery.
 
 ### Behavior Changes
 * Disallow the combination of DBOptions.use_direct_io_for_flush_and_compaction == true and DBOptions.writable_file_max_buffer_size == 0. This combination can cause WritableFileWriter::Append() to loop forever, and it does not make much sense in direct IO.